joi validation done

+++ controller/tutorials.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const Tutorial = require("../models/tutorials");
 const routes = require("../models/tutorials");
+const validator=require("../helper/joischemas");
 
 
 exports.getTutorial = (req, res) => {
@@ -31,7 +32,7 @@ exports.getSortedTutorial = (req, res) => {
    //console.log("hello",req.body);
     const sort={updatedAt: -1};
     const tutorial =Tutorial.find().sort(sort).then(tutorial => {
-        if(tutorial==null){
+        if(tutorial==null || tutorial==""){
             res.send("Tutorial Not Found");
         }else{
             res.json({
@@ -43,47 +44,61 @@ exports.getSortedTutorial = (req, res) => {
         console.log(err);
     })
 };
-exports.postTutorial = (req, res) => {
-    //res.send("hello");
-    //console.log(req.body); 
-    try{
-        let tutorial =new Tutorial(req.body);
-        tutorial.save().then((result)=>{
-        res.status(200).json({
-            result:result
-        });
-    })
-    }catch(err){
-        res.send(err);
-    }
+exports.postTutorial = async(req, res,next) => {
+   try{
+        const resultvalidated= await validator.swaggerschemasPOST.validateAsync(req.body)
+        // console.log(result.body);
+        let tutorial =new Tutorial(resultvalidated);
+        tutorial.save().then((results)=>{
+            res.status(200).json({
+                result:results
+            });
+        })
+   }catch(error){
+        res.status(422).json(error.message);
+   }
 };
 exports.putTutorial=async(req,res)=>{
     try{
-        let id=req.params.id;
-    const tutorial =await Tutorial.findByIdAndUpdate(id,{
-        title:req.body.title,
-        description:req.body.description,
-        published:req.body.published
-    },{new:true}).then((result)=>{
-        if(result==null){
-            res.send("Tutorial Not Found");
-        }else{
-            res.json({
-                tutorial:result
-            });
-        }
-    });    
+       let id = req.params.id.match(/^[0-9a-fA-F]{24}$/);
+       if (id==null){
+            throw new Error("check your id");
+       }else{
+            const resultBody= await validator.swaggerschemasPUT.validateAsync(req.body);
+            const tutorial =await Tutorial.findByIdAndUpdate(id,{
+                title:resultBody.title,
+                description:resultBody.description,
+                published:resultBody.published
+            },{new:true}).then((result)=>{
+                if(result==null || tutorial==""){
+                    res.send("Tutorial Not Found");
+                }else{
+                    res.json({
+                        tutorial:result
+                    });
+                }
+            });    
+       }
+         
     }catch(error){
-        res.status(400).send(error.message);
+        if(error.isJoi){
+            res.status(422).send(error.message);
+        }else{
+            res.status(422).send(error.message);
+        }
+        
     }
     
 
 };
 exports.deleteTutorial=(req,res)=>{
    try{
-    let id=req.params.id;
+    let id = req.params.id.match(/^[0-9a-fA-F]{24}$/);
+    if (id==null){
+         throw new Error("check your id");
+    }
     const tutorial = Tutorial.findByIdAndRemove(id).then((result)=>{
-        if(result==null){
+        if(result==null || tutorial==""){
             res.send("Tutorial Not Found");
         }else{
             res.json({
@@ -93,16 +108,19 @@ exports.deleteTutorial=(req,res)=>{
     });
     
    }catch(error){
-    res.status(400).send(error.message);
+    res.status(422).send(error.message);
    }
     
 }
 exports.findTutorial=async(req, res)=>{
     try{
-        let id=req.params.id;
+        let id = req.params.id.match(/^[0-9a-fA-F]{24}$/);
+       if (id==null){
+            throw new Error("check your id");
+       }
         //res.send(id)
         const tutorial=await Tutorial.findById(id).then(tutorial => {
-            if(tutorial==null){
+            if(tutorial==null || tutorial=="" ){
                 res.send("Tutorial Not Found");
             }else{
                 res.json({
@@ -112,7 +130,7 @@ exports.findTutorial=async(req, res)=>{
             
         });
     }catch(error){
-        res.status(400).send(error.message);
+        res.status(422).send(error.message);
     }
 }
 exports.findByTitleTutorial=async(req, res)=>{
@@ -132,16 +150,17 @@ exports.findByTitleTutorial=async(req, res)=>{
             field={"updatedAt":sorting}
         }
         const tutorial=await Tutorial.find({title:title}).sort(field).then(tutorial => {
-            if(tutorial==null){
+            if(tutorial==null || tutorial==""){
                 res.send("Tutorial Not Found");
             }else{
                 res.json({
                     tutorial:tutorial
                 });
             }
+            // console.log();
             
         });
     }catch(error){
-        res.status(400).send(error.message);
+        res.status(422).send(error.message);
     }
 }
